Feature/payload mismatch (#137)
* Feature/session expiration (#24)

* Bugfix/page type issue (#22)

* Update tests.yml

added test run on iOS 17

* Update tests.yml

cmd syntax issue resolved

* Bugfix/remove exclude from launch time (#20)

* Remove exclude property from LaunchTime Payload

* Remove exclude property from LaunchTime Payload

* Added log to BTTWebViewTracker

* Updated sdk memory warning message

* Added verifySessionStitching method to BTTWebViewTracker

* Updated verifySessionStitching method to BTTWebViewTracker

* Added a DEBUG condition to the verifyStitchingSession function so that it is only applicable in debug mode.

* Added siteId verification on WebViewTracker

* Added siteId verification on WebViewTracker

* Updated Verify session stitching error message and ReadMe.

* Added trafficSegmentName field to BTTimer

---------


* Set pageType empty on exclude-20- and error.rcv requests

* update version

---------


* Added Session Expiretion feature

* Add validation to restitchWebView method

* updated session manager

* Added _session static varrible to BlueTriangle

* added sessionExpiryDuration to configuration

* updated sessionExpiryDuration modifier

* Rename some session manager function name

* Fixed unit test exicution isssue

* Updated Version

* Remove commented code

---------


* Added Device model to payload NativeApp (#25)

* Added Device model to payload NativeApp

* Fixed unit test case issue

---------


* Feature/device model (#26)

* Added Device model to payload NativeApp

* Fixed unit test case issue

* Updated device model func

---------


* Feature/network protocol (#27)

* Implemented Network Protocol feature

* Fixed Unit test issue

* Fixed Watch OS unit test case issue

* Updated sdk version 3.11.0

* Updated sdk version 3.11.0

* Fixed cutom timer net State value issue

---------


* deprecated CustomCategories,CustomNumbers, CustomVariables class (#28)


* Feature/dynamic configuration (#29)

* Added dynamic config feature for sample rate

* Added ignoreViewControllers documentation readme

* Pointed remote config from local to live

* Remote config value change only on new session

* Added -FullSampleRate launch argument

* Added code to sample rate from session manager

* Added synchronize config to Config repo

* Fixed dynamic config unit test compie issue

* move config updater to session manager

* Added logger to remote config

* Added FULL_SAMPLE_RATE_ARGUMENT, NEW_SESSION_ON_LAUNCH_ARGUMENT to contants

* Updated dynamic configuration test cases

* Added code to observe config changes to config repository

* Fixed unit test compile issue

* Update remote confic url and Config class according to json

* Added Test url and updatePeriod for testing

* Update remote config url and update remote config agains siteId

* Added RemoteConfigAckReporter to acknowledge remote config change and error

* Added Error Type to remote config Ack

* Fixed divide by zero issue

* Added error log to ack reporter and updated

* Added error message to remote config ack

* Updated remote config error message

* Fixed capturedRequestCollector assignment issue

* Updated version and prod url

* Added local url for testing

* remove uncommented code

* Added Troubleshooting section for Launch Arguments to ReadMe

* Formated read me

* Formated read me

---------


* Feature/ignore vc with hot fix (#32)

* Feature/dynamic configuration ignore vc 


* Unit test bug (#33)

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Fix unit test issue

* Updated change log and version

---------


* Feature/dissable sdk (#34)

* Added Remaining feature tests in SDK Example App

* Added enable and dissable SDK feature functionality

* Added Code to dissable crash tracking feature

* Updated configuration field and update unit tests

* Disable WebView tracker on disable mode

* Updated some function name

* Added comment on some blue triangle function

* Stop web view tracking on disable sdk

* Updated code review changes

* Added code review changes

* Added reset session for disable and enable mode

* Updated and stated and stop All tracker methods

* Remove all cache on disable mode

* Change with test url

* Remove isSDKEnabled from SessionData class

* Fixed unit test exicution issue

* Clear all cache on disable mode

* Fixed launch argument issue

* Fixed optional session issue on each feature

* Fixed network capture optional session issue

* Fixed network capture optional session issue

* Updated remore url to production

* Remove unwanted code used for testing

* Fixed Upload Tests test case exicution isssue

* Fixed Web view enableAllTracking issue on disable mode

* update remote url to prod

* Fixed session issue found in testing

* Change remote config base url

* Remove Remote config chapter from readMe

* Fixed unit test issue

---------


* Feature/dissable sdk (#35)

* Added Remaining feature tests in SDK Example App

* Added enable and dissable SDK feature functionality

* Added Code to dissable crash tracking feature

* Updated configuration field and update unit tests

* Disable WebView tracker on disable mode

* Updated some function name

* Added comment on some blue triangle function

* Stop web view tracking on disable sdk

* Updated code review changes

* Added code review changes

* Added reset session for disable and enable mode

* Updated and stated and stop All tracker methods

* Remove all cache on disable mode

* Change with test url

* Remove isSDKEnabled from SessionData class

* Fixed unit test exicution issue

* Clear all cache on disable mode

* Fixed launch argument issue

* Fixed optional session issue on each feature

* Fixed network capture optional session issue

* Fixed network capture optional session issue

* Updated remore url to production

* Remove unwanted code used for testing

* Fixed Upload Tests test case exicution isssue

* Fixed Web view enableAllTracking issue on disable mode

* update remote url to prod

* Fixed session issue found in testing

* Change remote config base url

* Remove Remote config chapter from readMe

* Fixed unit test issue

* remove caching from url session configuration

---------


* Feature/clarity integration (#37)

* Added clarity support

* Update sdk version

* Fixed unit test case issue

* Updated read me

* Update clarity connector logger

* Update clarity connector logger

* Updated readme and changelog

* Updated code review changes

---------


* Feature/clarity integration (#38)

* Added clarity support

* Update sdk version

* Fixed unit test case issue

* Updated read me

* Update clarity connector logger

* Update clarity connector logger

* Updated readme and changelog

* Updated code review changes

* Update clarity connector method name

---------


* Updated remote config url, sample rate data types

* Updated eventType to entryType enum

* Update change log

---------

+++ Sources/BlueTriangle/DynamicConfig/BTTSavedRemoteConfig.swift
@@ -12,8 +12,8 @@ import Foundation
 class BTTSavedRemoteConfig: BTTRemoteConfig {
     var dateSaved: Millisecond
 
-    init(networkSampleRateSDK: Int?,
-         groupedViewSampleRate: Int?,
+    init(networkSampleRateSDK: Double?,
+         groupedViewSampleRate: Double?,
          enableRemoteConfigAck : Bool?,
          enableAllTracking : Bool?,
          enableScreenTracking: Bool?,
